requirements.txt
langchain
langchain_community
spotipy
crewai
python-dotenv
SpeechRecognition
PyAudio
gTTS
playsound
flask
flask-vite
flask-cors
tools

openai
duckduckgo-search
